Split up numba.targets.builtins
