+++ numba/targets/ufunc_db.py
@@ -47,116 +47,116 @@ def _fill_ufunc_db(ufunc_db):
     # some of these imports would cause a problem of circular
     # imports if done at global scope when importing the numba
     # module.
-    from . import builtins, npyfuncs, mathimpl, cmathimpl
+    from . import numbers, npyfuncs, mathimpl, cmathimpl
     from numba import numpy_support
 
     v = numpy_support.version
 
     ufunc_db[np.negative] = {
-        '?->?': builtins.int_invert_impl,
-        'b->b': builtins.int_negate_impl,
-        'B->B': builtins.int_negate_impl,
-        'h->h': builtins.int_negate_impl,
-        'H->H': builtins.int_negate_impl,
-        'i->i': builtins.int_negate_impl,
-        'I->I': builtins.int_negate_impl,
-        'l->l': builtins.int_negate_impl,
-        'L->L': builtins.int_negate_impl,
-        'q->q': builtins.int_negate_impl,
-        'Q->Q': builtins.int_negate_impl,
-        'f->f': builtins.real_negate_impl,
-        'd->d': builtins.real_negate_impl,
-        'F->F': builtins.complex_negate_impl,
-        'D->D': builtins.complex_negate_impl,
+        '?->?': numbers.int_invert_impl,
+        'b->b': numbers.int_negate_impl,
+        'B->B': numbers.int_negate_impl,
+        'h->h': numbers.int_negate_impl,
+        'H->H': numbers.int_negate_impl,
+        'i->i': numbers.int_negate_impl,
+        'I->I': numbers.int_negate_impl,
+        'l->l': numbers.int_negate_impl,
+        'L->L': numbers.int_negate_impl,
+        'q->q': numbers.int_negate_impl,
+        'Q->Q': numbers.int_negate_impl,
+        'f->f': numbers.real_negate_impl,
+        'd->d': numbers.real_negate_impl,
+        'F->F': numbers.complex_negate_impl,
+        'D->D': numbers.complex_negate_impl,
     }
 
     ufunc_db[np.absolute] = {
-        '?->?': builtins.int_abs_impl,
-        'b->b': builtins.int_abs_impl,
-        'B->B': builtins.uint_abs_impl,
-        'h->h': builtins.int_abs_impl,
-        'H->H': builtins.uint_abs_impl,
-        'i->i': builtins.int_abs_impl,
-        'I->I': builtins.uint_abs_impl,
-        'l->l': builtins.int_abs_impl,
-        'L->L': builtins.uint_abs_impl,
-        'q->q': builtins.int_abs_impl,
-        'Q->Q': builtins.uint_abs_impl,
-        'f->f': builtins.real_abs_impl,
-        'd->d': builtins.real_abs_impl,
-        'F->f': builtins.complex_abs_impl,
-        'D->d': builtins.complex_abs_impl,
+        '?->?': numbers.int_abs_impl,
+        'b->b': numbers.int_abs_impl,
+        'B->B': numbers.uint_abs_impl,
+        'h->h': numbers.int_abs_impl,
+        'H->H': numbers.uint_abs_impl,
+        'i->i': numbers.int_abs_impl,
+        'I->I': numbers.uint_abs_impl,
+        'l->l': numbers.int_abs_impl,
+        'L->L': numbers.uint_abs_impl,
+        'q->q': numbers.int_abs_impl,
+        'Q->Q': numbers.uint_abs_impl,
+        'f->f': numbers.real_abs_impl,
+        'd->d': numbers.real_abs_impl,
+        'F->f': numbers.complex_abs_impl,
+        'D->d': numbers.complex_abs_impl,
     }
 
     ufunc_db[np.sign] = {
-        'b->b': builtins.int_sign_impl,
-        'B->B': builtins.int_sign_impl,
-        'h->h': builtins.int_sign_impl,
-        'H->H': builtins.int_sign_impl,
-        'i->i': builtins.int_sign_impl,
-        'I->I': builtins.int_sign_impl,
-        'l->l': builtins.int_sign_impl,
-        'L->L': builtins.int_sign_impl,
-        'q->q': builtins.int_sign_impl,
-        'Q->Q': builtins.int_sign_impl,
-        'f->f': builtins.real_sign_impl,
-        'd->d': builtins.real_sign_impl,
+        'b->b': numbers.int_sign_impl,
+        'B->B': numbers.int_sign_impl,
+        'h->h': numbers.int_sign_impl,
+        'H->H': numbers.int_sign_impl,
+        'i->i': numbers.int_sign_impl,
+        'I->I': numbers.int_sign_impl,
+        'l->l': numbers.int_sign_impl,
+        'L->L': numbers.int_sign_impl,
+        'q->q': numbers.int_sign_impl,
+        'Q->Q': numbers.int_sign_impl,
+        'f->f': numbers.real_sign_impl,
+        'd->d': numbers.real_sign_impl,
         'F->F': npyfuncs.np_complex_sign_impl,
         'D->D': npyfuncs.np_complex_sign_impl,
     }
 
     ufunc_db[np.add] = {
-        '??->?': builtins.int_or_impl,
-        'bb->b': builtins.int_add_impl,
-        'BB->B': builtins.int_add_impl,
-        'hh->h': builtins.int_add_impl,
-        'HH->H': builtins.int_add_impl,
-        'ii->i': builtins.int_add_impl,
-        'II->I': builtins.int_add_impl,
-        'll->l': builtins.int_add_impl,
-        'LL->L': builtins.int_add_impl,
-        'qq->q': builtins.int_add_impl,
-        'QQ->Q': builtins.int_add_impl,
-        'ff->f': builtins.real_add_impl,
-        'dd->d': builtins.real_add_impl,
-        'FF->F': builtins.complex_add_impl,
-        'DD->D': builtins.complex_add_impl,
+        '??->?': numbers.int_or_impl,
+        'bb->b': numbers.int_add_impl,
+        'BB->B': numbers.int_add_impl,
+        'hh->h': numbers.int_add_impl,
+        'HH->H': numbers.int_add_impl,
+        'ii->i': numbers.int_add_impl,
+        'II->I': numbers.int_add_impl,
+        'll->l': numbers.int_add_impl,
+        'LL->L': numbers.int_add_impl,
+        'qq->q': numbers.int_add_impl,
+        'QQ->Q': numbers.int_add_impl,
+        'ff->f': numbers.real_add_impl,
+        'dd->d': numbers.real_add_impl,
+        'FF->F': numbers.complex_add_impl,
+        'DD->D': numbers.complex_add_impl,
     }
 
     ufunc_db[np.subtract] = {
-        '??->?': builtins.int_xor_impl,
-        'bb->b': builtins.int_sub_impl,
-        'BB->B': builtins.int_sub_impl,
-        'hh->h': builtins.int_sub_impl,
-        'HH->H': builtins.int_sub_impl,
-        'ii->i': builtins.int_sub_impl,
-        'II->I': builtins.int_sub_impl,
-        'll->l': builtins.int_sub_impl,
-        'LL->L': builtins.int_sub_impl,
-        'qq->q': builtins.int_sub_impl,
-        'QQ->Q': builtins.int_sub_impl,
-        'ff->f': builtins.real_sub_impl,
-        'dd->d': builtins.real_sub_impl,
-        'FF->F': builtins.complex_sub_impl,
-        'DD->D': builtins.complex_sub_impl,
+        '??->?': numbers.int_xor_impl,
+        'bb->b': numbers.int_sub_impl,
+        'BB->B': numbers.int_sub_impl,
+        'hh->h': numbers.int_sub_impl,
+        'HH->H': numbers.int_sub_impl,
+        'ii->i': numbers.int_sub_impl,
+        'II->I': numbers.int_sub_impl,
+        'll->l': numbers.int_sub_impl,
+        'LL->L': numbers.int_sub_impl,
+        'qq->q': numbers.int_sub_impl,
+        'QQ->Q': numbers.int_sub_impl,
+        'ff->f': numbers.real_sub_impl,
+        'dd->d': numbers.real_sub_impl,
+        'FF->F': numbers.complex_sub_impl,
+        'DD->D': numbers.complex_sub_impl,
     }
 
     ufunc_db[np.multiply] = {
-        '??->?': builtins.int_and_impl,
-        'bb->b': builtins.int_mul_impl,
-        'BB->B': builtins.int_mul_impl,
-        'hh->h': builtins.int_mul_impl,
-        'HH->H': builtins.int_mul_impl,
-        'ii->i': builtins.int_mul_impl,
-        'II->I': builtins.int_mul_impl,
-        'll->l': builtins.int_mul_impl,
-        'LL->L': builtins.int_mul_impl,
-        'qq->q': builtins.int_mul_impl,
-        'QQ->Q': builtins.int_mul_impl,
-        'ff->f': builtins.real_mul_impl,
-        'dd->d': builtins.real_mul_impl,
-        'FF->F': builtins.complex_mul_impl,
-        'DD->D': builtins.complex_mul_impl,
+        '??->?': numbers.int_and_impl,
+        'bb->b': numbers.int_mul_impl,
+        'BB->B': numbers.int_mul_impl,
+        'hh->h': numbers.int_mul_impl,
+        'HH->H': numbers.int_mul_impl,
+        'ii->i': numbers.int_mul_impl,
+        'II->I': numbers.int_mul_impl,
+        'll->l': numbers.int_mul_impl,
+        'LL->L': numbers.int_mul_impl,
+        'qq->q': numbers.int_mul_impl,
+        'QQ->Q': numbers.int_mul_impl,
+        'ff->f': numbers.real_mul_impl,
+        'dd->d': numbers.real_mul_impl,
+        'FF->F': numbers.complex_mul_impl,
+        'DD->D': numbers.complex_mul_impl,
     }
 
     if np.divide != np.true_divide:
@@ -276,20 +276,20 @@ def _fill_ufunc_db(ufunc_db):
     }
 
     ufunc_db[np.conjugate] = {
-        'b->b': builtins.real_conjugate_impl,
-        'B->B': builtins.real_conjugate_impl,
-        'h->h': builtins.real_conjugate_impl,
-        'H->H': builtins.real_conjugate_impl,
-        'i->i': builtins.real_conjugate_impl,
-        'I->I': builtins.real_conjugate_impl,
-        'l->l': builtins.real_conjugate_impl,
-        'L->L': builtins.real_conjugate_impl,
-        'q->q': builtins.real_conjugate_impl,
-        'Q->Q': builtins.real_conjugate_impl,
-        'f->f': builtins.real_conjugate_impl,
-        'd->d': builtins.real_conjugate_impl,
-        'F->F': builtins.complex_conjugate_impl,
-        'D->D': builtins.complex_conjugate_impl,
+        'b->b': numbers.real_conjugate_impl,
+        'B->B': numbers.real_conjugate_impl,
+        'h->h': numbers.real_conjugate_impl,
+        'H->H': numbers.real_conjugate_impl,
+        'i->i': numbers.real_conjugate_impl,
+        'I->I': numbers.real_conjugate_impl,
+        'l->l': numbers.real_conjugate_impl,
+        'L->L': numbers.real_conjugate_impl,
+        'q->q': numbers.real_conjugate_impl,
+        'Q->Q': numbers.real_conjugate_impl,
+        'f->f': numbers.real_conjugate_impl,
+        'd->d': numbers.real_conjugate_impl,
+        'F->F': numbers.complex_conjugate_impl,
+        'D->D': numbers.complex_conjugate_impl,
     }
 
     ufunc_db[np.exp] = {
@@ -522,109 +522,109 @@ def _fill_ufunc_db(ufunc_db):
 
     # logical ufuncs
     ufunc_db[np.greater] = {
-        '??->?': builtins.int_ugt_impl,
-        'bb->?': builtins.int_sgt_impl,
-        'BB->?': builtins.int_ugt_impl,
-        'hh->?': builtins.int_sgt_impl,
-        'HH->?': builtins.int_ugt_impl,
-        'ii->?': builtins.int_sgt_impl,
-        'II->?': builtins.int_ugt_impl,
-        'll->?': builtins.int_sgt_impl,
-        'LL->?': builtins.int_ugt_impl,
-        'qq->?': builtins.int_sgt_impl,
-        'QQ->?': builtins.int_ugt_impl,
-        'ff->?': builtins.real_gt_impl,
-        'dd->?': builtins.real_gt_impl,
+        '??->?': numbers.int_ugt_impl,
+        'bb->?': numbers.int_sgt_impl,
+        'BB->?': numbers.int_ugt_impl,
+        'hh->?': numbers.int_sgt_impl,
+        'HH->?': numbers.int_ugt_impl,
+        'ii->?': numbers.int_sgt_impl,
+        'II->?': numbers.int_ugt_impl,
+        'll->?': numbers.int_sgt_impl,
+        'LL->?': numbers.int_ugt_impl,
+        'qq->?': numbers.int_sgt_impl,
+        'QQ->?': numbers.int_ugt_impl,
+        'ff->?': numbers.real_gt_impl,
+        'dd->?': numbers.real_gt_impl,
         'FF->?': npyfuncs.np_complex_gt_impl,
         'DD->?': npyfuncs.np_complex_gt_impl,
     }
 
     ufunc_db[np.greater_equal] = {
-        '??->?': builtins.int_uge_impl,
-        'bb->?': builtins.int_sge_impl,
-        'BB->?': builtins.int_uge_impl,
-        'hh->?': builtins.int_sge_impl,
-        'HH->?': builtins.int_uge_impl,
-        'ii->?': builtins.int_sge_impl,
-        'II->?': builtins.int_uge_impl,
-        'll->?': builtins.int_sge_impl,
-        'LL->?': builtins.int_uge_impl,
-        'qq->?': builtins.int_sge_impl,
-        'QQ->?': builtins.int_uge_impl,
-        'ff->?': builtins.real_ge_impl,
-        'dd->?': builtins.real_ge_impl,
+        '??->?': numbers.int_uge_impl,
+        'bb->?': numbers.int_sge_impl,
+        'BB->?': numbers.int_uge_impl,
+        'hh->?': numbers.int_sge_impl,
+        'HH->?': numbers.int_uge_impl,
+        'ii->?': numbers.int_sge_impl,
+        'II->?': numbers.int_uge_impl,
+        'll->?': numbers.int_sge_impl,
+        'LL->?': numbers.int_uge_impl,
+        'qq->?': numbers.int_sge_impl,
+        'QQ->?': numbers.int_uge_impl,
+        'ff->?': numbers.real_ge_impl,
+        'dd->?': numbers.real_ge_impl,
         'FF->?': npyfuncs.np_complex_ge_impl,
         'DD->?': npyfuncs.np_complex_ge_impl,
     }
 
     ufunc_db[np.less] = {
-        '??->?': builtins.int_ult_impl,
-        'bb->?': builtins.int_slt_impl,
-        'BB->?': builtins.int_ult_impl,
-        'hh->?': builtins.int_slt_impl,
-        'HH->?': builtins.int_ult_impl,
-        'ii->?': builtins.int_slt_impl,
-        'II->?': builtins.int_ult_impl,
-        'll->?': builtins.int_slt_impl,
-        'LL->?': builtins.int_ult_impl,
-        'qq->?': builtins.int_slt_impl,
-        'QQ->?': builtins.int_ult_impl,
-        'ff->?': builtins.real_lt_impl,
-        'dd->?': builtins.real_lt_impl,
+        '??->?': numbers.int_ult_impl,
+        'bb->?': numbers.int_slt_impl,
+        'BB->?': numbers.int_ult_impl,
+        'hh->?': numbers.int_slt_impl,
+        'HH->?': numbers.int_ult_impl,
+        'ii->?': numbers.int_slt_impl,
+        'II->?': numbers.int_ult_impl,
+        'll->?': numbers.int_slt_impl,
+        'LL->?': numbers.int_ult_impl,
+        'qq->?': numbers.int_slt_impl,
+        'QQ->?': numbers.int_ult_impl,
+        'ff->?': numbers.real_lt_impl,
+        'dd->?': numbers.real_lt_impl,
         'FF->?': npyfuncs.np_complex_lt_impl,
         'DD->?': npyfuncs.np_complex_lt_impl,
     }
 
     ufunc_db[np.less_equal] = {
-        '??->?': builtins.int_ule_impl,
-        'bb->?': builtins.int_sle_impl,
-        'BB->?': builtins.int_ule_impl,
-        'hh->?': builtins.int_sle_impl,
-        'HH->?': builtins.int_ule_impl,
-        'ii->?': builtins.int_sle_impl,
-        'II->?': builtins.int_ule_impl,
-        'll->?': builtins.int_sle_impl,
-        'LL->?': builtins.int_ule_impl,
-        'qq->?': builtins.int_sle_impl,
-        'QQ->?': builtins.int_ule_impl,
-        'ff->?': builtins.real_le_impl,
-        'dd->?': builtins.real_le_impl,
+        '??->?': numbers.int_ule_impl,
+        'bb->?': numbers.int_sle_impl,
+        'BB->?': numbers.int_ule_impl,
+        'hh->?': numbers.int_sle_impl,
+        'HH->?': numbers.int_ule_impl,
+        'ii->?': numbers.int_sle_impl,
+        'II->?': numbers.int_ule_impl,
+        'll->?': numbers.int_sle_impl,
+        'LL->?': numbers.int_ule_impl,
+        'qq->?': numbers.int_sle_impl,
+        'QQ->?': numbers.int_ule_impl,
+        'ff->?': numbers.real_le_impl,
+        'dd->?': numbers.real_le_impl,
         'FF->?': npyfuncs.np_complex_le_impl,
         'DD->?': npyfuncs.np_complex_le_impl,
     }
 
     ufunc_db[np.not_equal] = {
-        '??->?': builtins.int_ne_impl,
-        'bb->?': builtins.int_ne_impl,
-        'BB->?': builtins.int_ne_impl,
-        'hh->?': builtins.int_ne_impl,
-        'HH->?': builtins.int_ne_impl,
-        'ii->?': builtins.int_ne_impl,
-        'II->?': builtins.int_ne_impl,
-        'll->?': builtins.int_ne_impl,
-        'LL->?': builtins.int_ne_impl,
-        'qq->?': builtins.int_ne_impl,
-        'QQ->?': builtins.int_ne_impl,
-        'ff->?': builtins.real_ne_impl,
-        'dd->?': builtins.real_ne_impl,
+        '??->?': numbers.int_ne_impl,
+        'bb->?': numbers.int_ne_impl,
+        'BB->?': numbers.int_ne_impl,
+        'hh->?': numbers.int_ne_impl,
+        'HH->?': numbers.int_ne_impl,
+        'ii->?': numbers.int_ne_impl,
+        'II->?': numbers.int_ne_impl,
+        'll->?': numbers.int_ne_impl,
+        'LL->?': numbers.int_ne_impl,
+        'qq->?': numbers.int_ne_impl,
+        'QQ->?': numbers.int_ne_impl,
+        'ff->?': numbers.real_ne_impl,
+        'dd->?': numbers.real_ne_impl,
         'FF->?': npyfuncs.np_complex_ne_impl,
         'DD->?': npyfuncs.np_complex_ne_impl,
     }
 
     ufunc_db[np.equal] = {
-        '??->?': builtins.int_eq_impl,
-        'bb->?': builtins.int_eq_impl,
-        'BB->?': builtins.int_eq_impl,
-        'hh->?': builtins.int_eq_impl,
-        'HH->?': builtins.int_eq_impl,
-        'ii->?': builtins.int_eq_impl,
-        'II->?': builtins.int_eq_impl,
-        'll->?': builtins.int_eq_impl,
-        'LL->?': builtins.int_eq_impl,
-        'qq->?': builtins.int_eq_impl,
-        'QQ->?': builtins.int_eq_impl,
-        'ff->?': builtins.real_eq_impl,
-        'dd->?': builtins.real_eq_impl,
+        '??->?': numbers.int_eq_impl,
+        'bb->?': numbers.int_eq_impl,
+        'BB->?': numbers.int_eq_impl,
+        'hh->?': numbers.int_eq_impl,
+        'HH->?': numbers.int_eq_impl,
+        'ii->?': numbers.int_eq_impl,
+        'II->?': numbers.int_eq_impl,
+        'll->?': numbers.int_eq_impl,
+        'LL->?': numbers.int_eq_impl,
+        'qq->?': numbers.int_eq_impl,
+        'QQ->?': numbers.int_eq_impl,
+        'ff->?': numbers.real_eq_impl,
+        'dd->?': numbers.real_eq_impl,
         'FF->?': npyfuncs.np_complex_eq_impl,
         'DD->?': npyfuncs.np_complex_eq_impl,
     }
@@ -824,85 +824,85 @@ def _fill_ufunc_db(ufunc_db):
 
     # bit twiddling functions
     ufunc_db[np.bitwise_and] = {
-        '??->?': builtins.int_and_impl,
-        'bb->b': builtins.int_and_impl,
-        'BB->B': builtins.int_and_impl,
-        'hh->h': builtins.int_and_impl,
-        'HH->H': builtins.int_and_impl,
-        'ii->i': builtins.int_and_impl,
-        'II->I': builtins.int_and_impl,
-        'll->l': builtins.int_and_impl,
-        'LL->L': builtins.int_and_impl,
-        'qq->q': builtins.int_and_impl,
-        'QQ->Q': builtins.int_and_impl,
+        '??->?': numbers.int_and_impl,
+        'bb->b': numbers.int_and_impl,
+        'BB->B': numbers.int_and_impl,
+        'hh->h': numbers.int_and_impl,
+        'HH->H': numbers.int_and_impl,
+        'ii->i': numbers.int_and_impl,
+        'II->I': numbers.int_and_impl,
+        'll->l': numbers.int_and_impl,
+        'LL->L': numbers.int_and_impl,
+        'qq->q': numbers.int_and_impl,
+        'QQ->Q': numbers.int_and_impl,
     }
 
     ufunc_db[np.bitwise_or] = {
-        '??->?': builtins.int_or_impl,
-        'bb->b': builtins.int_or_impl,
-        'BB->B': builtins.int_or_impl,
-        'hh->h': builtins.int_or_impl,
-        'HH->H': builtins.int_or_impl,
-        'ii->i': builtins.int_or_impl,
-        'II->I': builtins.int_or_impl,
-        'll->l': builtins.int_or_impl,
-        'LL->L': builtins.int_or_impl,
-        'qq->q': builtins.int_or_impl,
-        'QQ->Q': builtins.int_or_impl,
+        '??->?': numbers.int_or_impl,
+        'bb->b': numbers.int_or_impl,
+        'BB->B': numbers.int_or_impl,
+        'hh->h': numbers.int_or_impl,
+        'HH->H': numbers.int_or_impl,
+        'ii->i': numbers.int_or_impl,
+        'II->I': numbers.int_or_impl,
+        'll->l': numbers.int_or_impl,
+        'LL->L': numbers.int_or_impl,
+        'qq->q': numbers.int_or_impl,
+        'QQ->Q': numbers.int_or_impl,
     }
 
     ufunc_db[np.bitwise_xor] = {
-        '??->?': builtins.int_xor_impl,
-        'bb->b': builtins.int_xor_impl,
-        'BB->B': builtins.int_xor_impl,
-        'hh->h': builtins.int_xor_impl,
-        'HH->H': builtins.int_xor_impl,
-        'ii->i': builtins.int_xor_impl,
-        'II->I': builtins.int_xor_impl,
-        'll->l': builtins.int_xor_impl,
-        'LL->L': builtins.int_xor_impl,
-        'qq->q': builtins.int_xor_impl,
-        'QQ->Q': builtins.int_xor_impl,
+        '??->?': numbers.int_xor_impl,
+        'bb->b': numbers.int_xor_impl,
+        'BB->B': numbers.int_xor_impl,
+        'hh->h': numbers.int_xor_impl,
+        'HH->H': numbers.int_xor_impl,
+        'ii->i': numbers.int_xor_impl,
+        'II->I': numbers.int_xor_impl,
+        'll->l': numbers.int_xor_impl,
+        'LL->L': numbers.int_xor_impl,
+        'qq->q': numbers.int_xor_impl,
+        'QQ->Q': numbers.int_xor_impl,
     }
 
     ufunc_db[np.invert] = { # aka np.bitwise_not
-        '?->?': builtins.int_invert_impl,
-        'b->b': builtins.int_invert_impl,
-        'B->B': builtins.int_invert_impl,
-        'h->h': builtins.int_invert_impl,
-        'H->H': builtins.int_invert_impl,
-        'i->i': builtins.int_invert_impl,
-        'I->I': builtins.int_invert_impl,
-        'l->l': builtins.int_invert_impl,
-        'L->L': builtins.int_invert_impl,
-        'q->q': builtins.int_invert_impl,
-        'Q->Q': builtins.int_invert_impl,
+        '?->?': numbers.int_invert_impl,
+        'b->b': numbers.int_invert_impl,
+        'B->B': numbers.int_invert_impl,
+        'h->h': numbers.int_invert_impl,
+        'H->H': numbers.int_invert_impl,
+        'i->i': numbers.int_invert_impl,
+        'I->I': numbers.int_invert_impl,
+        'l->l': numbers.int_invert_impl,
+        'L->L': numbers.int_invert_impl,
+        'q->q': numbers.int_invert_impl,
+        'Q->Q': numbers.int_invert_impl,
     }
 
     ufunc_db[np.left_shift] = {
-        'bb->b': builtins.int_shl_impl,
-        'BB->B': builtins.int_shl_impl,
-        'hh->h': builtins.int_shl_impl,
-        'HH->H': builtins.int_shl_impl,
-        'ii->i': builtins.int_shl_impl,
-        'II->I': builtins.int_shl_impl,
-        'll->l': builtins.int_shl_impl,
-        'LL->L': builtins.int_shl_impl,
-        'qq->q': builtins.int_shl_impl,
-        'QQ->Q': builtins.int_shl_impl,
+        'bb->b': numbers.int_shl_impl,
+        'BB->B': numbers.int_shl_impl,
+        'hh->h': numbers.int_shl_impl,
+        'HH->H': numbers.int_shl_impl,
+        'ii->i': numbers.int_shl_impl,
+        'II->I': numbers.int_shl_impl,
+        'll->l': numbers.int_shl_impl,
+        'LL->L': numbers.int_shl_impl,
+        'qq->q': numbers.int_shl_impl,
+        'QQ->Q': numbers.int_shl_impl,
     }
 
     ufunc_db[np.right_shift] = {
-        'bb->b': builtins.int_shr_impl,
-        'BB->B': builtins.int_shr_impl,
-        'hh->h': builtins.int_shr_impl,
-        'HH->H': builtins.int_shr_impl,
-        'ii->i': builtins.int_shr_impl,
-        'II->I': builtins.int_shr_impl,
-        'll->l': builtins.int_shr_impl,
-        'LL->L': builtins.int_shr_impl,
-        'qq->q': builtins.int_shr_impl,
-        'QQ->Q': builtins.int_shr_impl,
+        'bb->b': numbers.int_shr_impl,
+        'BB->B': numbers.int_shr_impl,
+        'hh->h': numbers.int_shr_impl,
+        'HH->H': numbers.int_shr_impl,
+        'ii->i': numbers.int_shr_impl,
+        'II->I': numbers.int_shr_impl,
+        'll->l': numbers.int_shr_impl,
+        'LL->L': numbers.int_shr_impl,
+        'qq->q': numbers.int_shr_impl,
+        'QQ->Q': numbers.int_shr_impl,
     }
 
     # Inject datetime64 support
